ServerManager::handleClientEvent -> major logic to process epoll is almost done

+++ include/Webserv.hpp
@@ -90,7 +90,7 @@ std::string to_string(size_t value);
 
 uint64_t 	validateGetMbs(std::string param);
 
-std::string generateErrorPage(size_t status_code);
+std::string generateErrorPage(int status_code);
 
 //Debug
 class ServerConfig;
